FIX: data 타입 postgres에 맞게 변경

+++ backend/src/core/score/Score.entity.ts
@@ -22,13 +22,13 @@ export class Score extends Base {
   win: string;
 
   @Column({
-    type: 'tinyint',
+    type: 'smallint',
     comment: '점수',
   })
   score: number;
 
   @Column({
-    type: 'tinyint',
+    type: 'smallint',
     comment: '래더 점수(총 점수)',
   })
   LadderScore: number;
